+++ src/main/scala/org/kneelawk/learningopengl/buffers/GLArrayBufferSetOperation.scala
@@ -7,7 +7,7 @@ import org.lwjgl.system.MemoryUtil
 /**
  * Object representing a set operation.
  */
-case class GLArrayBufferSetOperation(offset: Int, bufLen: Int, bufData: Long) {
+case class GLArrayBufferSetOperation(offset: Long, bufLen: Long, bufData: Long) {
 
   if (offset < 0)
     throw new IllegalArgumentException("The offset cannot be negative")
@@ -18,25 +18,25 @@ case class GLArrayBufferSetOperation(offset: Int, bufLen: Int, bufData: Long) {
   /**
    * ByteBuffer version of the constructor.
    */
-  def this(offset: Int, buf: ByteBuffer) =
+  def this(offset: Long, buf: ByteBuffer) =
     this(offset, buf.remaining(), MemoryUtil.memAddress(buf))
 
   /**
    * ShortBuffer version of the constructor.
    */
-  def this(offset: Int, buf: ShortBuffer) =
+  def this(offset: Long, buf: ShortBuffer) =
     this(offset, buf.remaining() << 1, MemoryUtil.memAddress(buf))
 
   /**
    * IntBuffer version of the constructor.
    */
-  def this(offset: Int, buf: IntBuffer) =
+  def this(offset: Long, buf: IntBuffer) =
     this(offset, buf.remaining() << 2, MemoryUtil.memAddress(buf))
 
   /**
    * LongBuffer version of the constructor.
    */
-  def this(offset: Int, buf: LongBuffer) =
+  def this(offset: Long, buf: LongBuffer) =
     this(offset, buf.remaining() << 3, MemoryUtil.memAddress(buf))
 
   /**
@@ -48,6 +48,6 @@ case class GLArrayBufferSetOperation(offset: Int, bufLen: Int, bufData: Long) {
   /**
    * DoubleBuffer version of the constructor.
    */
-  def this(offset: Int, buf: DoubleBuffer) =
+  def this(offset: Long, buf: DoubleBuffer) =
     this(offset, buf.remaining() << 3, MemoryUtil.memAddress(buf))
 }
